init: do_mounts: Fix pointer warnings
../init/do_mounts.c:1157:19: warning: cast to smaller integer type 'int' from 'void *' [-Wvoid-pointer-to-int-cast]
        int thread_num = (int)unused;
                         ^~~~~~~~~~~
../init/do_mounts.c:1296:32: warning: cast to 'void *' from smaller integer type 'int' [-Wint-to-void-pointer-cast]
                kthread_run(prefetch_thread, (void *)index, name);
                                             ^~~~~~~~~~~~~
../include/linux/kthread.h:46:30: note: expanded from macro 'kthread_run'
                = kthread_create(threadfn, data, namefmt, ## __VA_ARGS__); \
                                           ^~~~
../include/linux/kthread.h:26:35: note: expanded from macro 'kthread_create'
        kthread_create_on_node(threadfn, data, NUMA_NO_NODE, namefmt, ##arg)

+++ init/do_mounts.c
@@ -1154,7 +1154,7 @@ static inline void prefetch_read_dir(void *name, char *buf, int len)
 static int prefetch_thread(void *unused)
 {
 	struct cpumask cpumask;
-	int thread_num = (int)unused;
+	int thread_num = (long)unused;
 	char *buf;
 	int index;
 	char *name;
@@ -1293,7 +1293,7 @@ out:
 
 	for (index = 0; index < PREFETCH_NUM; index++) {
 		snprintf(name, 16, "prefetch%d", index);
-		kthread_run(prefetch_thread, (void *)index, name);
+		kthread_run(prefetch_thread, (void *)(long)index, name);
 	}
 	}
 	first_time = 0;
